+++ doc/source/changelog.rst
@@ -3,6 +3,20 @@
 ChangeLog
 =========
 
+Version 4.3.0
+-------------
+
+This version of SOXS contains new features.
+
+* A new version of the spectral model used in the
+  :class:`~soxs.thermal_spectra.CloudyCIEGenerator` class has been provided, with
+  improved energy resolution. See :ref:`cloudy-spectra` for more details.
+* A new version of the spectral model used in the
+  :class:`~soxs.thermal_spectra.IGMGenerator` class has been provided, with
+  improved energy resolution. See :ref:`igm-spectra` for more details.
+* A new function to download table files for the thermal spectra models has been
+  provided. See :ref:`downloading-thermal-tables` for more details.
+
 Version 4.2.1
 -------------